splitting lens config helper types into their own module
